+++ src/main/java/de/maxhenkel/voicechat/voice/client/Client.java
@@ -118,6 +118,10 @@ public class Client extends Thread {
 
                     audioChannels.values().stream().filter(AudioChannel::canKill).forEach(AudioChannel::closeAndKill);
                     audioChannels.entrySet().removeIf(entry -> entry.getValue().isClosed());
+                } else if (in.getPacket() instanceof PingPacket) {
+                    PingPacket packet = (PingPacket) in.getPacket();
+                    Main.LOGGER.debug("Received ping {}, sending pong...", packet.getId());
+                    new NetworkMessage(packet, secret).sendToServer(this);
                 }
 
             }