+++ src/main/java/de/maxhenkel/voicechat/voice/server/ServerVoiceEvents.java
@@ -9,6 +9,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.event.server.FMLServerStartedEvent;
 import net.minecraftforge.fml.network.PacketDistributor;
 
+import javax.annotation.Nullable;
 import java.util.UUID;
 
 public class ServerVoiceEvents {
@@ -56,5 +57,11 @@ public class ServerVoiceEvents {
             Main.LOGGER.info("Disconnecting client " + player.getDisplayName().getString());
         }
     }
+
+    @Nullable
+    public Server getServer() {
+        return server;
+    }
+
 
 }